Changed File Structure. Plan to move teleop code from LocalizationTest to TeleOp tmrv.
