--- TeamCode/MeepMeep/build.gradle
@@ -1,16 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-}
-
-repositories {
-    maven { url = 'https://maven.brott.dev/' }
-}
-
-dependencies {
-    implementation 'com.acmerobotics.roadrunner:MeepMeep:0.1.6'
-}
