+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/TeleOp.java
@@ -1,0 +1,11 @@
+package org.firstinspires.ftc.teamcode.teleop;
+
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name ="Ouroboros")
+
+public class TeleOp extends LinearOpMode {
+
+    //add code
+
+}
