+++ settings.gradle
@@ -2,3 +2,4 @@ include ':FtcRobotController'
 include ':TeamCode'
 include ':MeepMeep'
 include ':Java'
+include ':MeepMeep:mylibrary'
